Server backup can also handle stopped domains

+++ VM/backup-restore/backup.sh
@@ -10,10 +10,10 @@ DOMAIN=$1
 BACKUPFOLDER="${2:-$PWD/$DOMAIN}"
 
 # Get the target disk
-TARGETS=$(virsh domblklist "$DOMAIN" --details | grep disk | awk '{print $3}')
+TARGETS=$(virsh domblklist "$DOMAIN" --details | awk '{print $2,$3}' | grep '^disk ' | awk '{print $2}')
 
 # Get the image page
-IMAGES=$(virsh domblklist "$DOMAIN" --details | grep disk | awk '{print $4}')
+IMAGES=$(virsh domblklist "$DOMAIN" --details | awk '{print $2,$4}' | grep '^disk ' | awk '{print $2}')
 
 # Create the snapshot/disk specification
 DISKSPEC=""
@@ -22,7 +22,10 @@ for TARGET in $TARGETS; do
 	DISKSPEC="$DISKSPEC --diskspec $TARGET,snapshot=external"
 done
 
-virsh snapshot-create-as --domain $DOMAIN --name "backup-$DOMAIN" --no-metadata --atomic --disk-only $DISKSPEC 
+STATE=$(virsh dominfo "$DOMAIN" | grep -F 'State: ' | cut -d':' -f2- | sed -E 's/^ +//')
+if [[ "$STATE" != "shut off" ]]; then
+  virsh snapshot-create-as --domain "$DOMAIN" --name "backup-$DOMAIN" --no-metadata --atomic --disk-only "$DISKSPEC"
+fi
 
 ## DANGER WILL ROBINSON, YOU NEED TO SET THE DISKS BACK IF THIS FAILS
 # Copy disk image
@@ -36,24 +39,23 @@ for IMAGE in $IMAGES; do
 	else 
 		rsync --archive --human-readable --whole-file --ignore-existing --sparse --progress "$IMAGE" "${BACKUPFOLDER}/${NAME}"
 	fi
-	
 done
 
+if [[ "$STATE" != "shut off" ]]; then
+  #Get the names of the backup images
+  BACKUPIMAGES=$(virsh domblklist "$DOMAIN" --details | awk '{print $2,$4}' | grep '^disk ' | awk '{print $2}')
 
-#Get the names of the backup images
-BACKUPIMAGES=$(virsh domblklist "$DOMAIN" --details | grep disk | awk '{print $4}')
+  # Merge changes back
+  for TARGET in $TARGETS; do
+    virsh blockcommit "$DOMAIN" "$TARGET" --wait --active --pivot --verbose
+  done
+  ## DANGER PAST
 
-# Merge changes back
-for TARGET in $TARGETS; do
-	virsh blockcommit "$DOMAIN" "$TARGET" --wait --active --pivot --verbose
-done
-
-## DANGER PAST
-
-# Cleanup left over backups
-for BACKUP in $BACKUPIMAGES; do
-	rm -f -- "$BACKUP"
-done
+  # Cleanup left over backups
+  for BACKUP in $BACKUPIMAGES; do
+    rm -f -- "$BACKUP"
+  done
+fi
 
 # Dump the configuration information.
 virsh dumpxml "$DOMAIN" > "$BACKUPFOLDER/$DOMAIN.xml"
